Distance_Sensor_Physical: Use environment
This makes the constructors of both classes similar and moves methods
that the mission monitor expects to the superclass. Make use of the
environment setup in the main script and tests.
